refact: extracts interfaces to their own file
