+++ app/src/hooks/api/useDataRequestApi.ts
@@ -1,24 +1,5 @@
 import { AxiosInstance } from 'axios';
-
-export interface CreateDataRequestPayload {
-  reason: string;
-  team_id?: string;
-}
-
-export interface DataRequestStatus {
-  data_request_status_id: string;
-  data_request_id: string;
-  comment_id: string | null;
-  request_status: 'REQUESTED' | 'APPROVED' | 'DENIED';
-}
-
-export interface DataRequestResponse {
-  data_request_id: string;
-  reason: string;
-  team_id: string;
-  requested_by: number;
-  data_request_status: DataRequestStatus;
-}
+import { CreateDataRequestPayload, DataRequestResponse } from 'interfaces/useDataRequestApi.interface';
 
 /**
  * Returns a set of supported api methods for working with data requests.
